+++ SharePointFramework/ProjectExtensions/src/models/ProjectTemplate.ts
@@ -6,6 +6,8 @@ import { Schema } from 'sp-js-provisioning'
 export interface IProjectTemplateSPItem {
   Id?: number
   IsDefaultTemplate?: boolean
+  IsDefaultExtensionsLocked?: boolean
+  IsDefaultListContentLocked?: boolean
   IconName?: string
   ListContentConfigLookupId?: number[]
   File?: { UniqueId: string; Name: string; Title: string; ServerRelativeUrl: string }
@@ -28,10 +30,12 @@ export class ProjectTemplate implements IDropdownOption {
   public text: string
   public subText: string
   public isDefault: boolean
+  public isDefaultExtensionsLocked: boolean
+  public isDefaultListContentLocked: boolean
   public iconName: string
-  public serverRelativeUrl: string
   public listContentConfigIds: number[]
   public projectTemplateId: number
+  public projectTemplateUrl: string
   public listExtensionIds: number[]
   public projectContentType: string
   public projectStatusContentType: string
@@ -48,6 +52,8 @@ export class ProjectTemplate implements IDropdownOption {
     this.text = spItem.FieldValuesAsText.Title
     this.subText = spItem.FieldValuesAsText.GtDescription
     this.isDefault = spItem?.IsDefaultTemplate
+    this.isDefaultExtensionsLocked = spItem?.IsDefaultExtensionsLocked
+    this.isDefaultListContentLocked = spItem?.IsDefaultListContentLocked
     this.iconName = spItem.IconName
     this.listContentConfigIds =
       spItem.ListContentConfigLookupId && spItem.ListContentConfigLookupId.length > 0
@@ -56,32 +62,33 @@ export class ProjectTemplate implements IDropdownOption {
     this.isProgram = spItem.GtIsProgram
     this.isParentProject = spItem.GtIsParentProject
     this.isHidden = spItem.IsHiddenTemplate
+      spItem.ListContentConfigLookupId?.length > 0 ? spItem.ListContentConfigLookupId : null
     this.projectTemplateId = spItem.GtProjectTemplateId
     this.listExtensionIds =
-      spItem.GtProjectExtensionsId && spItem.GtProjectExtensionsId.length > 0
-        ? spItem.GtProjectExtensionsId
-        : null
+      spItem.GtProjectExtensionsId?.length > 0 ? spItem.GtProjectExtensionsId : null
     this.projectContentType = spItem.GtProjectContentType
     this.projectStatusContentType = spItem.GtProjectStatusContentType
     this.projectColumns = spItem.GtProjectColumns
     this.projectCustomColumns = spItem.GtProjectCustomColumns
     this.projectPhaseTermId = spItem.GtProjectPhaseTermId
-    this.serverRelativeUrl = spItem.File?.ServerRelativeUrl
   }
 
   public async getSchema(): Promise<Schema> {
-    const schema = await this.web.getFileByServerRelativeUrl(this.serverRelativeUrl).getJSON()
-    if (!schema.Parameters){
-      schema.Parameters = {}
+    const schema = await this.web.getFileByServerRelativeUrl(this.projectTemplateUrl).getJSON()
+    schema.Parameters = schema.Parameters || {}
+    schema.Parameters.ProjectContentTypeId =
+      this?.projectContentType ?? schema.Parameters.ProjectContentTypeId
+    schema.Parameters.ProjectStatusContentTypeId =
+      this?.projectStatusContentType ?? schema.Parameters.ProjectStatusContentTypeId
+    schema.Parameters.ProvisionSiteFields =
+      this?.projectColumns ?? schema.Parameters.ProvisionSiteFields
+    schema.Parameters.CustomSiteFields =
+      this?.projectCustomColumns ?? schema.Parameters.CustomSiteFields
+    if (!schema.Parameters.TermSetIds) {
+      schema.Parameters.TermSetIds = {}
     }
-      schema.Parameters.ProjectContentTypeId = this?.projectContentType ?? schema.Parameters.ProjectContentTypeId 
-      schema.Parameters.ProjectStatusContentTypeId = this?.projectStatusContentType ?? schema.Parameters.ProjectStatusContentTypeId
-      schema.Parameters.ProvisionSiteFields = this?.projectColumns ?? schema.Parameters.ProvisionSiteFields
-      schema.Parameters.CustomSiteFields = this?.projectCustomColumns ?? schema.Parameters.CustomSiteFields
-      if (!schema.Parameters.TermSetIds) {
-        schema.Parameters.TermSetIds = {}
-      }
-      schema.Parameters.TermSetIds.GtProjectPhase = this?.projectPhaseTermId ?? schema.Parameters.TermSetIds.GtProjectPhase
+    schema.Parameters.TermSetIds.GtProjectPhase =
+      this?.projectPhaseTermId ?? schema.Parameters.TermSetIds.GtProjectPhase
     return schema
   }
 }