+++ SharePointFramework/ProjectExtensions/src/projectSetup/tasks/PreTask/index.ts
@@ -5,9 +5,7 @@ import { SpEntityPortalService } from 'sp-entityportal-service'
 import initSpfxJsom, { ExecuteJsomQuery } from 'spfx-jsom'
 import { BaseTask, BaseTaskError, IBaseTaskParams } from '../@BaseTask'
 import _ from 'underscore'
-import {ITaxonomySession, Session} from '@pnp/sp-taxonomy'
-
-
+import { ITaxonomySession, Session } from '@pnp/sp-taxonomy'
 
 export class PreTask extends BaseTask {
   constructor(data: IProjectSetupData) {
@@ -18,9 +16,7 @@ export class PreTask extends BaseTask {
    * Execute PreTask
    *
    * @param params Task parameters
-   * @param onProgress On progress function
    */
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   public async execute(params: IBaseTaskParams): Promise<IBaseTaskParams> {
     params.templateSchema = await this.data.selectedTemplate.getSchema()
     if (!params.properties.forceTemplate) {
@@ -47,35 +43,59 @@ export class PreTask extends BaseTask {
     }
   }
 
+  /**
+   * Checks that the parameters are valid
+   * 
+   * @param params - Task params
+   */
   private async validateParameters(params: IBaseTaskParams): Promise<void> {
-    const parametersToValidate: string[] = (_.toArray(params.templateSchema.Parameters) as string[])
+    const parametersToValidate: string[] = _.toArray(params.templateSchema.Parameters) as string[]
     const [termSetIds] = parametersToValidate.filter((param) => _.isObject(param))
-    const contentTypesToValidate = parametersToValidate.filter((param) => !_.isObject(param)).filter((ct) => ct.includes('0x'))
+    const contentTypesToValidate = parametersToValidate
+      .filter((param) => !_.isObject(param))
+      .filter((ct) => ct.includes('0x'))
     await this.validateTermSetIds(termSetIds)
     await this.validateContentTypes(contentTypesToValidate)
   }
 
+  /**
+   * Checks that the term set IDs are valid
+   * 
+   * @param termSetIds - Term set IDs
+   */
   private async validateTermSetIds(termSetIds: any) {
     const taxonomySession: ITaxonomySession = new Session()
-    const termSet = await taxonomySession.getDefaultSiteCollectionTermStore().getTermSetById(termSetIds.GtProjectPhase).get()
+    const termSet = await taxonomySession
+      .getDefaultSiteCollectionTermStore()
+      .getTermSetById(termSetIds.GtProjectPhase)
+      .get()
     if (!termSet.Name) {
       this.logError(`Failed to validate term set ${termSetIds.GtProjectPhase}`)
-      throw new BaseTaskError(this.taskName, strings.PreTaskTermSetIdValidationErrorMessage, strings.TermSetDoesNotExistError)
+      throw new BaseTaskError(
+        this.taskName,
+        strings.PreTaskTermSetIdValidationErrorMessage,
+        strings.TermSetDoesNotExistError
+      )
     }
   }
 
   /**
    * Checks that the content types are valid and exist in the hub
-   * @param params IBaseTaskParams
+   * 
+   * @param contentTypes - Content types to validate
    */
-  private async validateContentTypes(contenttypes: string[]): Promise<void> {
+  private async validateContentTypes(contentTypes: string[]): Promise<void> {
     await Promise.all(
-      contenttypes.map(async (ct) => {
+      contentTypes.map(async (ct) => {
         try {
           await this.data.hub.web.contentTypes.getById(ct).get()
         } catch (error) {
           this.logError(`Failed to validate content type ${ct}`)
-          throw new BaseTaskError(this.taskName, strings.PreTaskContentTypeValidationErrorMessage, error)
+          throw new BaseTaskError(
+            this.taskName,
+            strings.PreTaskContentTypeValidationErrorMessage,
+            error
+          )
         }
       })
     )